- На завершение quiz-a собрал модельку QuizResult и отправил в HistoryStorage
- Реализовал логику всех функций в QuizViewModel

- в методе ResultActivity.start() передавал id результата

- Подключил зависимости Room

- Верстка ResultActivity

+++ app/src/main/java/com/geektech/quizapp_gt_4_2/data/history/HistoryStorage.java
@@ -11,7 +11,7 @@ public class HistoryStorage implements IHistoryStorage {
 
 
     @Override
-    public int save(QuizResult result) {
+    public long save(QuizResult result) {
         return result.id;
     }
